// frontend/src/app/api/model/user-response-dto.ts
/**
 * Pharmacy API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserResponseDto { 
    id?: number;
    firstName?: string;
    lastName?: string;
    username?: string;
    role?: string;
}