// frontend/src/app/api/model/user-create-dto.ts
/**
 * Pharmacy API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserCreateDto { 
    firstName: string;
    lastName: string;
    position?: string;
    username: string;
    password: string;
    role?: string;
}